Added exception handling

+++ upload_pics.py
@@ -26,7 +26,11 @@ account = open('Account.txt','r')
 username = account.readline().rstrip()
 password = account.readline().rstrip()
 account.close()
-file = open('./img_urls.txt','r')
+try:
+    file = open('./img_urls.txt','r')
+except:
+    print("Web Scraping Failed")
+    exit(1)
 urls = file.readlines()
 urls = [i[:-1] for i in urls]
 urls = [i.split('^^') for i in urls]
@@ -34,26 +38,31 @@ file.close()
 
 cli = client(username,password)
 
-for i in urls:
-    r = requests.get(i[0])
-    f = open("image",'wb')
-    f.write(r.content)
-    viewer = subprocess.Popen(['eog','image'])
-    print("URL - " + i[2])
-    c = input("upload?[y/n]")
-    if c=='n':
+try:
+    for i in urls:
+        r = requests.get(i[0])
+        f = open("image",'wb')
+        f.write(r.content)
+        viewer = subprocess.Popen(['eog','image'])
+        print("URL - " + i[2])
+        c = input("upload?[y/n]")
+        if c=='n':
+            viewer.terminate()
+            viewer.kill()
+            os.remove('image')
+            f.close()
+            continue
+        Reformat_Image('image')
+        cli.upload('out.png',"Post mirrored from Reddit\nCREDITS: "+i[1]+".\n.\n.\n.\n.\n.\n#memes #dankmemes #reddit #sadmemes #spicymemes #humor #funny #pewdiepie #fortnite #dailymemes #thanosmemes #minecraft #memereview #lwiay #pewnews #pewpew #memestgram #memedaily #memesfunny #epicmemes #normiememes #spicymeme #stolenmemes #dankmemesdaily #mememachine")
         viewer.terminate()
         viewer.kill()
         os.remove('image')
+        os.remove('out.png')
         f.close()
-        continue
-    Reformat_Image('image')
-    cli.upload('out.png',"Post mirrored from Reddit\nCREDITS: "+i[1]+".\n.\n.\n.\n.\n.\n#memes #dankmemes #reddit #sadmemes #spicymemes #humor #funny #pewdiepie #fortnite #dailymemes #thanosmemes #minecraft #memereview #lwiay #pewnews #pewpew #memestgram #memedaily #memesfunny #epicmemes #normiememes #spicymeme #stolenmemes #dankmemesdaily #mememachine")
-    viewer.terminate()
-    viewer.kill()
-    os.remove('image')
-    os.remove('out.png')
-    f.close()
-
-os.remove('post_urls.txt')
-os.remove('img_urls.txt')
+finally:
+    if(os.path.exists("./image")):
+        os.remove("image")
+    if(os.path.exists("./out.png")):
+        os.remove("out.png")
+    os.remove('post_urls.txt')
+    os.remove('img_urls.txt')
